#23 fix binmode startup check and move dropdown
